Add files via upload
Objectives
Use simple imperative programming constructs (conditional statements and loops) to implement C
functions.
Getting Started
1. Depending on your operating system, download the starting code from Brightspace (lab2_win.zip or
lab2_macOS.zip) and extract the zipped file.
2. Open the unzipped folder and double-click on the workspace file. You are now ready to start working
in your lab.
3. Note that even though it is good practice to name the file where the main function is located main.c,
it is not mandatory. You can compile and run main_lab2.c as you did when you installed VS Code
in lab 1.
4. In main_lab2.c, locate these assignment statements at the top of the file:
/* Replace these lines with your information */
const char* author = "";
const char* student_number = "";
Replace the empty strings with strings containing your name and student number. For example, student
Jane Doe, who has student number 100111222, will edit the assignment statements to look like this:
const char* author = "Jane Doe";
const char* student_number = "100111222";
Do not change the variable names!
The “Rules”
For this lab, you may assume that the functions will be called with valid arguments.
SYSC 2006 - Foundations of Imperative Programming
Page 2 of 3
Test Driven Development
We are going to approach this lab using test-driven development.
(1) Read all the exercises below and ensure you understand each function's expected output for the
different cases.
(2) In the file main_lab2.c, locate the main function and write test cases for all your functions.
Some test cases for exercise (1) are provided as an example. Make sure that the test harness is
complete for all functions
(3) Compile and run main_lab2.c. All the tests should fail. If any test passes, it is just luck
because the answer matched the hard-coded value in the return statement.
(4) Solve exercises (1) to (4) inside the file lab2.c
Exercise 1: Alarm Clock
Write a function named alarm_clock. The function has two parameters. The first parameter is an
integer named day and encodes the day of the week: 0=Sun, 1=Mon, 2=Tue, ..., 6=Sat. The second
parameter is a Boolean named vacation, and it is true when we are on vacation. The function returns
an integer that indicates when the alarm will ring. On weekdays when we are not on vacation, the alarm
should ring at 7 a.m., i.e., the function should return 7. On weekends when we are not on vacation and
on weekdays when we are on vacation, the alarm should ring at 10 a.m., i.e., the function should return
10. On weekends when we are on vacation, the alarm should be off, i.e., the function should return -1.
Exercise 2: Count and Replace
Write a function named count_replace that takes a positive integer (greater than zero) as the input
parameter. The function should then count from 1 until the integer entered by the user, replacing
multiples of 3 with "Fizz", multiples of 5 with "Buzz", and multiples of both 3 and 5 with "FizzBuzz".
It displays the resulting sequence on the screen. The function does not return any value (i.e., it is void).
Example (no return value, just display in the screen):
>>> count_replace(16);
1 2 Fizz 4 Buzz Fizz 7 8 Fizz Buzz 11 Fizz 13 14 FizzBuzz 16
Exercise 3: Check Prime Number
Write a function named is_prime. It takes a positive integer as a parameter. The function checks if
the positive given number (greater than zero) is a prime number. The function returns a Boolean that is
true if the number is prime.
Exercise 4: Sum of even number in a range
Write a function named sum_even that takes a positive integer as a parameter. The function should
calculate and return the sum of all even numbers from 1 to the function’s input parameter.
Example:
int result = sumEven(10); /// should return 30